requirements.txt
pytorch
tensorboard
gym
tqdm